backend/app/services/ollama_service.py
import os
import json
import time
import base64
import requests
from typing import List, Dict, Any
from PIL import Image
import numpy as np
from app.core.config import settings


class OllamaService:
    """Ollama服务类"""
    
    def __init__(self):
        self.base_url = settings.ollama_base_url
        self.model_name = settings.ollama_model_name
    
    def _encode_image_to_base64(self, image_path: str) -> str:
        """将图片编码为base64"""
        try:
            with open(image_path, "rb") as image_file:
                return base64.b64encode(image_file.read()).decode('utf-8')
        except Exception as e:
            raise Exception(f"图片编码失败: {str(e)}")
    
    def _calculate_image_similarity(self, image1_path: str, image2_path: str) -> float:
        """计算两张图片的相似度（用于验证）"""
        try:
            # 加载图片
            img1 = Image.open(image1_path).convert('RGB')
            img2 = Image.open(image2_path).convert('RGB')
            
            # 统一尺寸
            size = (224, 224)
            img1 = img1.resize(size)
            img2 = img2.resize(size)
            
            # 转换为numpy数组
            arr1 = np.array(img1)
            arr2 = np.array(img2)
            
            # 计算均方误差
            mse = np.mean((arr1 - arr2) ** 2)
            
            # 转换为相似度分数 (0-1)
            max_mse = 255 ** 2
            similarity = 1 - (mse / max_mse)
            
            return max(0, min(1, similarity))
        except Exception as e:
            print(f"计算图片相似度失败: {str(e)}")
            return 0.5  # 默认值
    
    def _call_ollama_api(self, prompt: str, images: List[str]) -> Dict[str, Any]:
        """调用Ollama API"""
        url = f"{self.base_url}/api/generate"
        
        payload = {
            "model": self.model_name,
            "prompt": prompt,
            "images": images,
            "stream": False,
            "options": {
                "temperature": 0.1,
                "top_p": 0.9,
                "max_tokens": 2048  # 减少token数量以加快响应
            }
        }
        
        try:
            print(f"调用Ollama API: {url}")
            print(f"模型: {self.model_name}")
            print(f"图片数量: {len(images)}")
            
            response = requests.post(url, json=payload, timeout=600)  # 减少超时时间到30秒
            response.raise_for_status()
            
            result = response.json()
            print(f"Ollama API响应成功: {len(result.get('response', ''))} 字符")
            return result
            
        except requests.exceptions.Timeout:
            print("Ollama API调用超时，返回模拟数据")
            return self._get_mock_response(time.time())
        except requests.exceptions.ConnectionError:
            print("无法连接到Ollama服务，返回模拟数据")
            return self._get_mock_response(time.time())
        except requests.exceptions.RequestException as e:
            print(f"Ollama API调用失败: {str(e)}，返回模拟数据")
            return self._get_mock_response(time.time())
    
    def analyze_image_differences(self, image1_path: str, image2_path: str) -> Dict[str, Any]:
        """分析两张图片的差异"""
        start_time = time.time()
        
        try:
            print("=== Ollama服务开始分析 ===")
            
            # 首先计算图片相似度
            similarity_score = self._calculate_image_similarity(image1_path, image2_path)
            print(f"计算得到的相似度: {similarity_score:.4f}")
            
            # 提高阈值，只有在图片几乎完全相同时才跳过AI分析
            if similarity_score > 0.9995:  # 提高到0.9995
                print("图片几乎完全相同，跳过AI分析")
                return {
                    "similarity_score": similarity_score,
                    "differences": [],
                    "alert_level": "info",
                    "summary": "图片基本相同，未检测到显著差异",
                    "processing_time": time.time() - start_time
                }
            
            print("开始AI内容分析...")
            
            # 编码图片
            image1_base64 = self._encode_image_to_base64(image1_path)
            image2_base64 = self._encode_image_to_base64(image2_path)
            
            # 构建更敏感的分析提示词
            prompt = """
请非常仔细地分析这两张图片的差异。这是监控场景的图片对比分析，请特别注意：

**重点关注的变化（即使很小也要检测）：**
1. 人物：人的出现、消失、移动、姿势变化（即使只有部分身体可见）
2. 物体：物体的增加、减少、移动、状态变化
3. 设备：指示灯、显示屏、开关状态的变化
4. 环境：光线变化、阴影变化、背景变化
5. 细节：任何可见的细节差异，包括时间戳、文字等

**分析要求：**
- 即使差异很小，也要仔细检测并报告
- 特别注意人物出现这种重要的安全相关变化
- 对于检测到的差异，请提供准确的置信度
- 如果发现任何变化，请详细描述

请以JSON格式返回分析结果：
{
    "similarity_score": 0.85,
    "differences": [
        {
            "type": "person_detected",
            "description": "检测到人物出现",
            "confidence": 0.95,
            "bbox": [100, 150, 200, 250],
            "severity": "high"
        }
    ],
    "alert_level": "warning",
    "summary": "总体分析摘要"
}

请确保JSON格式正确，不要包含额外的文本。
"""
            
            print("发送AI分析请求...")
            # 调用API
            response = self._call_ollama_api(prompt, [image1_base64, image2_base64])
            print("收到AI响应")
            
            # 解析响应
            if 'response' not in response:
                raise Exception("API响应格式错误")
            
            print(f"AI原始响应: {response['response'][:200]}...")  # 只显示前200个字符
            
            # 清理响应文本，移除markdown代码块标记
            cleaned_response = response['response'].strip()
            if cleaned_response.startswith('```json'):
                cleaned_response = cleaned_response[7:]  # 移除 ```json
            if cleaned_response.startswith('```'):
                cleaned_response = cleaned_response[3:]  # 移除 ```
            if cleaned_response.endswith('```'):
                cleaned_response = cleaned_response[:-3]  # 移除结尾的 ```
            cleaned_response = cleaned_response.strip()
            
            print(f"清理后的响应: {cleaned_response[:200]}...")
            
            # 尝试解析JSON响应
            try:
                result = json.loads(cleaned_response)
                print("JSON解析成功")
                print(f"AI检测到的差异数量: {len(result.get('differences', []))}")
                
                # 使用计算得到的相似度，而不是AI返回的
                result['similarity_score'] = similarity_score
                
                # 降低过滤阈值，更敏感地检测差异
                if similarity_score > 0.99 and len(result.get('differences', [])) > 0:
                    print("检测到可能的误判，相似度很高但AI报告了差异")
                    # 过滤掉低置信度的差异，但降低阈值
                    filtered_differences = [
                        diff for diff in result.get('differences', [])
                        if diff.get('confidence', 0) > 0.7  # 降低置信度阈值
                    ]
                    result['differences'] = filtered_differences
                    print(f"过滤后差异数量: {len(filtered_differences)}")
                    
                    if len(filtered_differences) == 0:
                        result['alert_level'] = 'info'
                        result['summary'] = '图片基本相同，未检测到显著差异'
                
                result['processing_time'] = time.time() - start_time
                print(f"Ollama分析完成，耗时: {result['processing_time']:.2f}秒")
                return result
            except json.JSONDecodeError as e:
                print(f"JSON解析失败: {str(e)}")
                # 如果JSON解析失败，使用文本解析
                return self._parse_text_response(response['response'], similarity_score)
                
        except requests.exceptions.ConnectionError:
            print("Ollama连接失败，返回模拟数据")
            # 当Ollama服务不可用时，返回模拟数据
            return self._get_mock_response(start_time)
        except Exception as e:
            print(f"Ollama分析失败: {str(e)}")
            raise Exception(f"Ollama API调用失败: {str(e)}")
    
    def _parse_text_response(self, text: str, similarity_score: float) -> Dict[str, Any]:
        """解析文本响应（当JSON解析失败时使用）"""
        # 简单的文本解析逻辑
        result = {
            "similarity_score": similarity_score,
            "differences": [],
            "alert_level": "info",
            "summary": text[:200] + "..." if len(text) > 200 else text
        }
        
        # 根据关键词判断告警级别
        if any(keyword in text.lower() for keyword in ['错误', '故障', '异常', '危险']):
            result["alert_level"] = "error"
        elif any(keyword in text.lower() for keyword in ['警告', '注意', '变化']):
            result["alert_level"] = "warning"
        
        return result
    
    def _get_mock_response(self, start_time: float) -> Dict[str, Any]:
        """返回模拟的分析结果"""
        return {
            "similarity_score": 0.95,
            "differences": [],
            "alert_level": "info",
            "summary": "图片基本相同，未检测到显著差异",
            "processing_time": time.time() - start_time
        }
    
    def test_connection(self) -> bool:
        """测试Ollama服务连接"""
        try:
            url = f"{self.base_url}/api/tags"
            response = requests.get(url, timeout=10)
            response.raise_for_status()
            
            # 检查模型是否可用
            models = response.json().get('models', [])
            model_names = [model.get('name', '') for model in models]
            
            if self.model_name in model_names:
                print(f"Ollama服务连接正常，模型 {self.model_name} 可用")
                return True
            else:
                print(f"Ollama服务连接正常，但模型 {self.model_name} 不可用")
                print(f"可用模型: {model_names}")
                return False
                
        except Exception as e:
            print(f"Ollama服务连接失败: {str(e)}")
            return False


# 创建全局实例
ollama_service = OllamaService() 