backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
pydantic-settings==2.6.1
python-multipart==0.0.20
pillow==11.0.0
requests==2.32.3
python-dotenv==1.0.1
sqlalchemy==2.0.36
alembic==1.14.1
pytest==8.3.4
httpx==0.28.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
numpy==1.26.4 